cpu,decode: fix UB for invalid instructions
* ((s)->isa.inst.val) >> 32 is UB, where ubsan reports "shift exponent
  32 is too large for 32-bit type 'unsigned int'"

+++ include/cpu/decode.h
@@ -90,7 +90,7 @@ finish:
 #define INSTPAT(pattern, ...) do { \
   uint64_t key, mask, shift; \
   pattern_decode(pattern, STRLEN(pattern), &key, &mask, &shift); \
-  if (((INSTPAT_INST(s) >> shift) & mask) == key) { \
+  if ((((uint64_t)INSTPAT_INST(s) >> shift) & mask) == key) { \
     INSTPAT_MATCH(s, ##__VA_ARGS__); \
     goto *(__instpat_end); \
   } \
